# 智能土地.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "土地监测系统后端接口文档。\n用户可认领土地，通过线上支付发送施肥浇水请求，管理员端接收请求人工执行。\n",
    "version": "1.0.0"
  },
...
  "paths": {
    "/auth/wx-login": {
      "post": {
        "summary": "微信小程序登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "认证"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "wx.login()获取的临时登录凭证"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJIUzI1NiJ9..."
                            },
                            "tokenExpire": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "isNewUser": {
                              "type": "boolean"
                            },
                            "userInfo": {
                              "$ref": "#/components/schemas/UserInfo"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/auth/bind-phone": {
      "post": {
        "summary": "绑定手机号",
        "deprecated": false,
        "description": "",
        "tags": [
          "认证"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "getPhoneNumber获取的code"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "phone": {
                              "type": "string",
                              "example": "138****8888"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/auth/admin-login": {
      "post": {
        "summary": "管理员登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "认证"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "description": "MD5传输"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string"
                            },
                            "tokenExpire": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "adminInfo": {
                              "$ref": "#/components/schemas/AdminInfo"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/profile": {
      "get": {
        "summary": "获取用户信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDetail"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新用户信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "avatar": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/lands": {
      "get": {
        "summary": "获取我的认领土地列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageResultClaimLand"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/orders": {
      "get": {
        "summary": "获取我的订单列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 50
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "订单状态筛选",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_PAY",
                "PAID",
                "PROCESSING",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageResultOrder"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/orders": {
      "post": {
        "summary": "创建服务订单",
        "deprecated": false,
        "description": "",
        "tags": [
          "服务订单"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "landId",
                  "serviceType"
                ],
                "properties": {
                  "landId": {
                    "type": "integer"
                  },
                  "serviceType": {
                    "type": "string",
                    "enum": [
                      "WATER",
                      "FERTILIZE"
                    ],
                    "description": "WATER浇水 FERTILIZE施肥"
                  },
                  "quantity": {
                    "type": "integer",
                    "default": 1
                  },
                  "remark": {
                    "type": "string",
                    "maxLength": 200
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderDetail"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/orders/{orderId}": {
      "get": {
        "summary": "获取订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "服务订单"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderDetail"
                        }
                      }
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/orders/{orderId}/cancel": {
      "put": {
        "summary": "取消订单",
        "deprecated": false,
        "description": "仅限PENDING_PAY状态",
        "tags": [
          "服务订单"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "0成功，其他为错误码",
            "example": 0
          },
          "msg": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "example": 10001
          },
          "nickname": {
            "type": "string",
            "example": "微信用户"
          },
          "avatar": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserDetail": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer"
          },
          "nickname": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "claimCount": {
            "type": "integer",
            "description": "认领土地数"
          },
          "orderCount": {
            "type": "integer",
            "description": "订单数"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminInfo": {
        "type": "object",
        "properties": {
          "adminId": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN"
            ]
          }
        }
      },
...
      "ClaimLandItem": {
        "type": "object",
        "properties": {
          "claimId": {
            "type": "integer"
          },
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string"
          },
          "landImage": {
            "type": "string"
          },
          "area": {
            "type": "number",
            "format": "double"
          },
          "areaUnit": {
            "type": "string"
          },
          "claimTime": {
            "type": "string",
            "format": "date-time"
          },
          "expireTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          }
        }
      },
      "OrderDetail": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer"
          },
          "landId": {
            "type": "integer"
          },
          "landName": {
            "type": "string"
          },
          "serviceType": {
            "type": "string",
            "enum": [
              "WATER",
              "FERTILIZE"
            ]
          },
          "serviceTypeName": {
            "type": "string",
            "example": "浇水"
          },
          "quantity": {
            "type": "integer"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "remark": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_PAY",
              "PAID",
              "PROCESSING",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "statusName": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "payTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "acceptTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completeTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "operator": {
            "type": "object",
            "properties": {
              "adminId": {
                "type": "integer"
              },
              "adminName": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "completeImages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "completeRemark": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PageResultClaimLand": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClaimLandItem"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "PageResultOrder": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}